refactor(webhook): move stream candidates into ingestion/stream.rs
Gate the stream-candidate adapters once at the module instead of on each function.
